# [aarch64] speed up add and sub operators with overflow detection