@sebpop 2019-04-30
Further testing on aarch64 showed an error with this patch.  I am withdrawing this patch for now and I will resubmit it for review after I fix the problem.