@sebpop 2019-04-30
The problem was that the register `w2` that I have used as temporary was also used by the compiler.  I fixed the issue by using a temporary variable `tmp` assigned to a register `w5` that is now fixed and can be used in the inline assembly.  I will push an amended patch with the fix that passes `make test` on top of the inc/dec patch that enables the detection of `HAVE_ASM_GOTO`.