@nikic 2019-05-02
Your original code, which compiler is this generated with? Do you know whether it has __builtin_saddll_overflow?