@sebpop 2019-05-02
The original code was generated by gcc 7.3.0 on ubuntu.  Yes the compiler has the builtin, here are the perf numbers with the builtin:
```
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
BM_add_before               13.3 ns         13.3 ns     52617609
BM_add_after                4.80 ns         4.80 ns    145923156
BM_sub_before               8.72 ns         8.72 ns     80254220
BM_sub_after                4.80 ns         4.80 ns    145931253
BM_add_builtin_saddll       7.85 ns         7.85 ns     89180422
BM_add_builtin_ssubll       7.85 ns         7.85 ns     89188549
```