@dstogov 2019-05-07
Looks fine to me now.
I think, this should be merged into 7.4.
@nikic ?