@nikic 2019-05-07
Yeah, this looks good for 7.4.

@sebpop If it's not too hard, could you also show the assembly that __builtin_saddll_overflow generates? It looks like it is better than the fallback implementation, but worse than the hand-coded variant, and it would be interesting to see why.