@dstogov 2019-05-07
@nikic as I remember, __builtin_saddll_overflow() also was worse on x86[_64] GCC.