@dstogov 2019-05-07
Merged into PHP-7.4 and  above.