@sebpop 2019-05-07
Here is the code side by side of the builtin and the new inline asm: 
```
fun_builtin_saddll:             fun_asm:                    
        ldr     x2, [x2]                ldr    x5, [x1]     
        ldr     x1, [x1]                ldr    x6, [x2]     
        add     x3, x1, x2              adds    x5, x5, x6  
        eon     x4, x1, x2              bvs     .L10        
        eor     x5, x3, x2              mov     w6, 4       
        tst     x5, x4                  str     x5, [x0]    
        bmi     .L7                     str     w6, [x0, 8] 
        mov     w1, 4                   ret                 
        str     x3, [x0]        L10:                       
.L5:                                    ldr     x1, [x1]    
        str     w1, [x0, 8]             mov     w3, 5       
        ret                             ldr     x2, [x2]    
.L7:                                    str     w3, [x0, 8] 
        scvtf   d1, x1                  scvtf   d0, x1      
        scvtf   d0, x2                  scvtf   d1, x2      
        mov     w1, 5                   fadd    d0, d0, d1  
        fadd    d0, d1, d0              str     d0, [x0]    
        str     d0, [x0]                ret                 
        b       .L5
```
I used gcc (Ubuntu/Linaro 7.4.0-1ubuntu1~18.04) 7.4.0 on a Graviton A1 instance
compiling the following code at -O3:
```
#include "zend_strtod.h"

void fun_builtin_saddll(zval *result, zval *op1, zval *op2) {
  long long llresult;
  if (UNEXPECTED(__builtin_saddll_overflow(Z_LVAL_P(op1), Z_LVAL_P(op2), &llresult))) {
    ZVAL_DOUBLE(result, (double) Z_LVAL_P(op1) + (double) Z_LVAL_P(op2));
  } else {
    ZVAL_LONG(result, llresult);
  }
}

void fun_asm(zval *result, zval *op1, zval *op2) {
        __asm__ goto(
                "ldr    x5, [%1]\n\t"
                "ldr    x6, [%2]\n\t"
                "adds   x5, x5, x6\n\t"
                "bvs    %l5\n\t"
                "mov    w6, %3\n\t"
                "str    x5, [%0]\n\t"
                "str    w6, [%0, %c4]\n"
                :
                : "r"(&result->value),
                  "r"(&op1->value),
                  "r"(&op2->value),
                  "n"(IS_LONG),
                  "n"(ZVAL_OFFSETOF_TYPE)
                : "x5", "x6", "cc", "memory"
                : overflow);
        return;
 overflow: ZEND_ATTRIBUTE_COLD_LABEL
          ZVAL_DOUBLE(result, (double) Z_LVAL_P(op1) + (double) Z_LVAL_P(op2));
}
```