@nikic 2019-05-07
@sebpop Thanks! Sad to see that GCC still doesn't lower these correctly :(