@sebpop 2019-05-07
The code generated by clang-7 for the saddll builtin looks like the inline asm
```
fun_builtin_saddll:
	ldr	x8, [x1]
	ldr	x9, [x2]
	adds	x10, x8, x9
	b.vs	.LBB0_2
	orr	w8, wzr, #0x4
	str	x10, [x0]
	str	w8, [x0, #8]
	ret
.LBB0_2:
	scvtf	d0, x8
	scvtf	d1, x9
	fadd	d0, d0, d1
	mov	w8, #5
	str	d0, [x0]
	str	w8, [x0, #8]
	ret
[...]
	.ident	"clang version 7.0.0-3~ubuntu0.18.04.1 (tags/RELEASE_700/final)"
```