@sebpop 2019-05-08
gcc-10 trunk as of today generates good code for the builtin saddll on aarch64:
```
fun_builtin_saddll:
        ldr     x1, [x1]
        ldr     x2, [x2]
        adds    x3, x1, x2
        bvs     .L3
        mov     w1, 4
        str     x3, [x0]
.L5:
        str     w1, [x0, 8]
        ret
.L3:
        scvtf   d0, x1
        scvtf   d1, x2
        mov     w1, 5
        fadd    d0, d0, d1
        str     d0, [x0]
        b       .L5
```